refactor: modularize ExerciseClient
